requirements_database.txt
# KrushiBandhu AI - Database Dependencies
# Phase 1.1: Data Foundation & Automated Engine

# Database
mysql-connector-python==8.2.0
pymysql==1.1.0

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Environment Management
python-dotenv==1.0.0

# Logging and Utilities
colorlog==6.8.0

# API and HTTP
requests==2.31.0
aiohttp==3.9.1

# Date and Time
python-dateutil==2.8.2

# JSON and Data Serialization
orjson==3.9.10

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
